+++ collector/mongodb_collector.go
@@ -40,7 +40,7 @@ type MongodbCollectorOpts struct {
 	CollectTopMetrics        bool
 	CollectIndexUsageStats   bool
 	CollectConnPoolStats     bool
-	CollectShardingStatus    string
+	SuppressCollectShardingStatus    bool
 }
 
 func (in *MongodbCollectorOpts) toSessionOps() *shared.MongoSessionOpts {
@@ -227,7 +227,7 @@ func (exporter *MongodbCollector) collectMongos(client *mongo.Client, ch chan<- 
 		serverStatus.Export(ch)
 	}
 
-	if exporter.Opts.CollectShardingStatus == "true" {
+	if !exporter.Opts.SuppressCollectShardingStatus  {
 		log.Debug("Collecting Sharding Status")
 		shardingStatus := mongos.GetShardingStatus(client)
 		if shardingStatus != nil {