+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.11.2]
 ### Added
-- [PMM-6361](https://jira.percona.com/browse/PMM-6361): New flag `--collect.shardingstatus` can be used to disable Sharding Status. This flag is set to `true` by default. 
-On a large scale cluster it could help you to disable the all mongoS induced load to config tables. [@vrazvan-adobe](https://github.com/vrazvan-adobe)
+- [PMM-6361](https://jira.percona.com/browse/PMM-6361): New flag `--suppress.collectshardingstatus` can be used to disable the collection of Sharding Status. This flag is not set by default. 
+On a large scale cluster it could help you to disable the mongoS exporters induced load to config tables. [@vrazvan-adobe](https://github.com/vrazvan-adobe)
 
 ### Fixed
 - [PMM-6361](https://jira.percona.com/browse/PMM-6361): `runtime error: invalid memory address or nil pointer dereference`, source="sharding_status.go:166" .